Extract shared preview UI module
