+++ ProvisionQLCore/PreviewUI/Views/AppVersion.swift
@@ -9,9 +9,8 @@ import Foundation
 enum AppVersion {
     /// Returns the formatted version string for the app (e.g., "v1.0.0 (123)")
     static var versionString: String {
-        let bundle = Bundle(for: PreviewViewController.self)
-        let version = bundle.object(forInfoDictionaryKey: "CFBundleShortVersionString") as? String ?? "Unknown"
-        let build = bundle.object(forInfoDictionaryKey: "CFBundleVersion") as? String ?? "Unknown"
+        let version = Bundle.main.object(forInfoDictionaryKey: "CFBundleShortVersionString") as? String ?? "Unknown"
+        let build = Bundle.main.object(forInfoDictionaryKey: "CFBundleVersion") as? String ?? "Unknown"
         return "v\(version) (\(build))"
     }
 }
